requirements.txt
Django==3.2.24
djongo==1.3.6
sqlparse==0.2.4
pymongo<4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
python-dateutil>=2.8.2
Pillow>=10.1.0
